@@ -1,6 +1,6 @@
 package net.soundvibe.lasher.map.model;
 
-public class UnsafeAccess extends RuntimeException {
+public final class UnsafeAccess extends RuntimeException {
 
     public UnsafeAccess(Throwable cause) {
         super(cause);
