@@ -9,16 +9,22 @@ import java.nio.*;
 import java.nio.channels.FileChannel;
 import java.nio.file.*;
 import java.util.*;
+import java.util.concurrent.locks.ReadWriteLock;
+import java.util.concurrent.locks.ReentrantReadWriteLock;
 
 import static net.soundvibe.lasher.util.BytesSupport.BYTE_ORDER;
 
 public abstract class MemoryMapped implements Closeable {
 
-    private static final int CHUNK_SIZE = 128 * 1024 * 1024; //128 MB
+	private static final int MIN_CHUNK_SIZE = 32 * 1024 * 1024; // 32 MB
+	private static final int MAX_CHUNK_SIZE = 64 * 1024 * 1024; // 64 MB
+
+    private final int chunkSize;
     private final FileType fileType;
-    protected MappedByteBuffer[] buffers;
+    protected MappedBuffer[] buffers;
     protected long size;
     private final Path baseDir;
+    protected final ReadWriteLock rwLock;
 
     private static final Class<?> UNSAFE_CLASS = resolveUnsafeClass();
     private static final Unsafe UNSAFE = resolveUnsafe();
@@ -28,7 +34,9 @@ public abstract class MemoryMapped implements Closeable {
         Objects.requireNonNull(baseDir, "baseDir is null");
         this.fileType = fileType;
         this.baseDir = baseDir;
-        final long length = Math.max(CHUNK_SIZE, roundTo4096(defaultLength));
+        this.rwLock = new ReentrantReadWriteLock(true);
+        this.chunkSize = Math.max(MIN_CHUNK_SIZE, Math.min(MAX_CHUNK_SIZE, (int) roundTo4096(defaultLength)));
+        final long length = Math.max(this.chunkSize, roundTo4096(defaultLength));
 
         var fileStats = readFileStats(baseDir, fileType, length);
         this.size = Math.max(length, fileStats.totalSize);
@@ -41,20 +49,12 @@ public abstract class MemoryMapped implements Closeable {
     public abstract long getLong(long pos);
     public abstract void putLong(long pos, long val);
 
-    private static final class FileStats {
-        final long totalSize;
-        final MappedByteBuffer[] buffers;
-
-        private FileStats(long totalSize, MappedByteBuffer[] buffers) {
-            this.totalSize = totalSize;
-            this.buffers = buffers;
-        }
-    }
+	private record FileStats(long totalSize, MappedBuffer[] buffers) {}
 
     private FileStats readFileStats(final Path baseDir, FileType fileType, long defaultLength) {
         var path = baseDir.resolve(fileType.filename);
         if (Files.notExists(path)) {
-            return new FileStats(defaultLength, new MappedByteBuffer[0]);
+            return new FileStats(defaultLength, new MappedBuffer[0]);
         }
 
         var file = path.toFile();
@@ -64,12 +64,12 @@ public abstract class MemoryMapped implements Closeable {
             try (var rf = new RandomAccessFile(file, "rw");
                 var fc = rf.getChannel()) {
 
-                var totalBuffersSize = (int) Math.ceil(Math.max(1d, (double) fileSize / CHUNK_SIZE));
-                var totalBuffers = new MappedByteBuffer[totalBuffersSize];
+                var totalBuffersSize = (int) Math.ceil(Math.max(1d, (double) fileSize / chunkSize));
+                var totalBuffers =  new MappedBuffer[totalBuffersSize];
                 int index = 0;
-                for (long i = 0; i < fileSize; i+=CHUNK_SIZE) {
-                    totalBuffers[index] = fc.map(FileChannel.MapMode.READ_WRITE, i, CHUNK_SIZE);
-                    totalBuffers[index].order(BYTE_ORDER);
+                for (long i = 0; i < fileSize; i+= chunkSize) {
+                    totalBuffers[index] = new MappedBuffer(
+                            fc.map(FileChannel.MapMode.READ_WRITE, i, chunkSize), UNSAFE, ADDRESS_FIELD);
                     index++;
                 }
                 return new FileStats(fileSize, totalBuffers);
@@ -83,13 +83,22 @@ public abstract class MemoryMapped implements Closeable {
         return this.size;
     }
 
-    public void remap(long newLength) {
+    private void remap(long newLength) {
         var newSize = roundTo4096(newLength);
         mapAndResize(newSize);
         this.size = newSize;
     }
 
     public void doubleGrow() {
+        rwLock.writeLock().lock();
+        try {
+            doubleGrowNoLock();
+        } finally {
+            rwLock.writeLock().unlock();
+        }
+    }
+
+    public void doubleGrowNoLock() {
         remap(this.size * 2);
     }
 
@@ -97,8 +106,8 @@ public abstract class MemoryMapped implements Closeable {
     public void close() {
         for (var buffer : buffers) {
             if (buffer != null) {
-                buffer.force();
-                unmap(buffer);
+                buffer.flush();
+                buffer.close();
             }
         }
     }
@@ -106,22 +115,9 @@ public abstract class MemoryMapped implements Closeable {
     public void clear() {
         for (var buffer : buffers) {
             if (buffer != null) {
-                try {
-                    var addr = ADDRESS_FIELD.getLong(buffer);
-                    UNSAFE.setMemory(addr, buffer.capacity(), (byte) 0);
-                } catch (IllegalAccessException e) {
-                    throw new UnsafeAccess(e);
-                }
+                buffer.clear();
             }
         }
-    }
-
-    private static void unmap(ByteBuffer byteBuffer) {
-        if (byteBuffer == null || !byteBuffer.isDirect()) return;
-        if (UNSAFE == null) {
-            throw new UnsupportedOperationException("Unsafe not supported on this platform");
-        }
-        UNSAFE.invokeCleaner(byteBuffer);
     }
 
     private static Class<?> resolveUnsafeClass() {
@@ -166,14 +162,8 @@ public abstract class MemoryMapped implements Closeable {
         expandBuffers(bufferIndex, newSize);
     }
 
-    private MappedByteBuffer mapBuffer(int bufferIndex, int bufferSize, long fileSize) {
-        try (var f = new RandomAccessFile(baseDir.resolve(fileType.filename).toFile(), "rw");
-             var fc = f.getChannel()) {
-            if (f.length() < fileSize) {
-                f.setLength(fileSize);
-                this.size = fileSize;
-            }
-
+    private MappedByteBuffer mapBuffer(FileChannel fc, int bufferIndex, int bufferSize) {
+        try {
             var pos = resolveBufferPos(bufferIndex);
             var buffer = fc.map(FileChannel.MapMode.READ_WRITE, pos, bufferSize);
             buffer.order(BYTE_ORDER);
@@ -195,20 +185,20 @@ public abstract class MemoryMapped implements Closeable {
     }
 
     protected int convertPos(long absolutePos, int bufferIndex) {
-        long startPos = (long) CHUNK_SIZE * (long) bufferIndex;
+        long startPos = (long) chunkSize * (long) bufferIndex;
         int bufferPos = (int) (absolutePos - startPos);
-        if (bufferPos < 0 || bufferPos > CHUNK_SIZE) {
-            throw new IndexOutOfBoundsException("Buffer pos " + bufferPos + " is out of bounds: " + CHUNK_SIZE + " for pos: " + absolutePos + " and buffer index: " + bufferIndex);
+        if (bufferPos < 0 || bufferPos > chunkSize) {
+            throw new IndexOutOfBoundsException("Buffer pos " + bufferPos + " is out of bounds: " + chunkSize + " for pos: " + absolutePos + " and buffer index: " + bufferIndex);
         }
         return bufferPos;
     }
 
     protected int findBufferIndex(long pos) {
-        return (int) Math.floorDiv(pos, CHUNK_SIZE);
+        return (int) Math.floorDiv(pos, chunkSize);
     }
 
     protected int resolveBufferIndex(long pos) {
-        int ix = (int) Math.floorDiv(pos, CHUNK_SIZE);
+        int ix = (int) Math.floorDiv(pos, chunkSize);
         if (ix < 0 || ix >= buffers.length) {
             throw new IndexOutOfBoundsException("Buffer index " + ix + " is out of total length: " + buffers.length + " for pos " + pos);
         }
@@ -219,9 +209,22 @@ public abstract class MemoryMapped implements Closeable {
         if (newPartition + 1 > buffers.length) {
             int oldLength = buffers.length;
             buffers = Arrays.copyOf(buffers, newPartition + 1);
-            for (int i = oldLength; i < buffers.length; i++) {
-                unmap(buffers[i]);
-                buffers[i] = mapBuffer(i, CHUNK_SIZE, newSize);
+            try (var f = new RandomAccessFile(baseDir.resolve(fileType.filename).toFile(), "rw");
+                 var fc = f.getChannel()) {
+                if (f.length() < newSize) {
+                    f.setLength(newSize);
+                    this.size = newSize;
+                }
+
+                for (int i = oldLength; i < buffers.length; i++) {
+                    var buffer = buffers[i];
+                    if (buffer != null) {
+                        buffer.close();
+                    }
+                    buffers[i] = new MappedBuffer(mapBuffer(fc, i, chunkSize), UNSAFE, ADDRESS_FIELD);
+                }
+            } catch (IOException e) {
+                throw new UncheckedIOException(e);
             }
         }
     }
